Introduce views-based UI
