+++ src/ui/mod.rs
@@ -1,1 +1,5 @@
-pub mod style;
+mod style;
+mod views;
+mod app;
+
+pub use app::App;
